app/requirements.txt
spyne==2.14.0
suds-community==1.1.2
lxml==4.9.2
mysql-connector-python==8.0.32
pymongo==4.3.3
cryptography==40.0.1
pydantic==1.10.7
python-jose[cryptography]==3.3.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0
zeep==4.2.1
soaplib==2.0.0b2
flask==2.2.3
werkzeug==2.2.3